requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-environ==0.11.2
django-modeltranslation==0.19.3
django-rosetta==0.10.0
django-summernote==0.8.20.0
idna==3.7
pillow==10.3.0
polib==1.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
